Refactored to account for upload button changing from component to container. Fixes #98
