+++ client/src/components/Card/index.js
@@ -12,8 +12,8 @@ import {
   FolderIcon
 } from '../Icon'
 import InnerButton from '../InnerButton'
-import UploadButton from '../UploadButton'
 import FolderNavButton from '../FolderNavButton'
+import UploadButton from '../../containers/UploadButton'
 
 const Card = ({ children }) => (
   <div className={styles.cardHolder}>
@@ -26,7 +26,7 @@ export const UploadCard = ({ fileType, upload, uploadFolder }) => (
     <UploadButton
       name='Upload File'
       onClick={uploadFolder}
-      icon={<UploadIcon onClick={() => {}} />}
+      icon={<UploadIcon onClick={() => { }} />}
     />
   </div>
 )
@@ -54,12 +54,12 @@ export const FolderFunctionsCard = ({ createFolder, uploadFolder }) => (
       <FolderNavButton
         name='Create Folder'
         onClick={createFolder}
-        icon={<CreateFolderIcon onClick={() => {}} />}
+        icon={<CreateFolderIcon onClick={() => { }} />}
       />
       <FolderNavButton
         name='Upload Folder'
         onClick={uploadFolder}
-        icon={<UploadIcon onClick={() => {}} />}
+        icon={<UploadIcon onClick={() => { }} />}
       />
     </div>
   </div>
@@ -72,12 +72,12 @@ export const TrashCard = ({ name, fileType, id, deleteForever, restore }) => (
       <InnerButton
         name={`Delete ${fileType}`}
         onClick={() => deleteForever(id)}
-        icon={<TrashIcon onClick={() => {}} />}
+        icon={<TrashIcon onClick={() => { }} />}
       />
       <InnerButton
         name={`Restore ${fileType}`}
         onClick={() => restore(id)}
-        icon={<RestoreIcon onClick={() => {}} />}
+        icon={<RestoreIcon onClick={() => { }} />}
       />
     </div>
   </Card>
@@ -90,24 +90,24 @@ export const FileCard = ({
   downloadFile,
   moveFile
 }) => (
-  <Card>
-    <div className={styles.cardHeader}>
-      <div className={styles.cardTitle}>
-        {fileName}{' '}
-        <div className={styles.titleIconHolder}>
-          <FileIcon onClick={() => {}} />
+    <Card>
+      <div className={styles.cardHeader}>
+        <div className={styles.cardTitle}>
+          {fileName}{' '}
+          <div className={styles.titleIconHolder}>
+            <FileIcon onClick={() => { }} />
+          </div>
         </div>
       </div>
-    </div>
-    {/* <div className={styles.cardFooter}> */}
-    <div className={styles.fileCardBody}>
-      <TrashIcon onClick={() => trashFile(fileId)} />
-      <MoveIcon onClick={() => moveFile(fileId)} />
-      <DownloadIcon onClick={() => downloadFile(fileId)} />
-    </div>
-    {/* </div> */}
-  </Card>
-)
+      {/* <div className={styles.cardFooter}> */}
+      <div className={styles.fileCardBody}>
+        <TrashIcon onClick={() => trashFile(fileId)} />
+        <MoveIcon onClick={() => moveFile(fileId)} />
+        <DownloadIcon onClick={() => downloadFile(fileId)} />
+      </div>
+      {/* </div> */}
+    </Card>
+  )
 
 export const FolderCard = ({
   folderName,
@@ -116,28 +116,28 @@ export const FolderCard = ({
   downloadFolder,
   openFolder
 }) => (
-  <Card>
-    <div className={styles.cardTitle}>
-      {folderName}
-      <div className={styles.titleIconHolder}>
-        <FolderIcon onClick={() => openFolder(folderId)} />
+    <Card>
+      <div className={styles.cardTitle}>
+        {folderName}
+        <div className={styles.titleIconHolder}>
+          <FolderIcon onClick={() => openFolder(folderId)} />
+        </div>
       </div>
-    </div>
-    <div className={styles.folderText}>Open Folder</div>
+      <div className={styles.folderText}>Open Folder</div>
 
-    <div className={styles.folderCardBody} onClick={() => openFolder(folderId)}>
-      <TrashIcon
-        onClick={e => {
-          e.stopPropagation()
-          trashFolder(folderId)
-        }}
-      />
-      <DownloadIcon
-        onClick={e => {
-          e.stopPropagation()
-          downloadFolder(folderId)
-        }}
-      />
-    </div>
-  </Card>
-)
+      <div className={styles.folderCardBody} onClick={() => openFolder(folderId)}>
+        <TrashIcon
+          onClick={e => {
+            e.stopPropagation()
+            trashFolder(folderId)
+          }}
+        />
+        <DownloadIcon
+          onClick={e => {
+            e.stopPropagation()
+            downloadFolder(folderId)
+          }}
+        />
+      </div>
+    </Card>
+  )
